[Fix #161363372] Rewrite the tests Test the application behavoiur based on user role
