+++ .gitignore
@@ -3,3 +3,5 @@ node_modules/
 UI/spec
 UI/tests
 
+yarn.lock
+.DS_Store
